requirements.txt
jieba
matplotlib
nonebot-adapter-onebot
nonebot2
numpy
openai
pandas
scikit_learn
seaborn
websockets
wordcloud
